Move APIs to microeditor-envelope-protocol
